+++ changelog.md
@@ -1,5 +1,6 @@
 # 1.4.1
 
+- Generated objects can be selected and deleted by pressing the undo button
 - Camera zoom is accounted for when calculating gamemode bounds
 - Added a warning popup when starting the generator in platformer mode